pyproject.toml
[project]
name = "pcomp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "agate==1.7.1",
    "appdirs==1.4.4",
    "attrs==23.1.0",
    "babel==2.12.1",
    "cachetools==5.3.1",
    "certifi==2024.7.4",
    "cffi==1.15.1",
    "cfgv==3.4.0",
    "chardet==5.2.0",
    "charset-normalizer==3.2.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "dbt-bigquery==1.6.5",
    "dbt-core==1.6.15",
    "dbt-extractor==0.4.1",
    "dbt-semantic-interfaces==0.2.0",
    "diff-cover==7.7.0",
    "distlib==0.3.7",
    "fake-useragent>=2.2.0",
    "filelock==3.12.4",
    "google-api-core==2.12.0",
    "google-auth==2.23.0",
    "google-cloud-bigquery==3.11.4",
    "google-cloud-core==2.3.3",
    "google-cloud-dataproc==5.6.0",
    "google-cloud-storage==2.11.0",
    "google-crc32c==1.5.0",
    "google-resumable-media==2.6.0",
    "googleapis-common-protos==1.60.0",
    "grpc-google-iam-v1==0.12.6",
    "grpcio==1.58.0",
    "grpcio-status==1.58.0",
    "hologram==0.0.16",
    "identify==2.5.30",
    "idna==3.7",
    "importlib-metadata==6.8.0",
    "iniconfig==2.0.0",
    "isodate==0.6.1",
    "jinja2==3.1.4",
    "jinja2-simple-tags==0.5.0",
    "jsonschema==3.2.0",
    "leather==0.3.4",
    "logbook==1.5.3",
    "lxml>=6.0.2",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.3",
    "mashumaro==3.8.1",
    "mdurl==0.1.2",
    "minimal-snowplow-tracker==0.0.2",
    "more-itertools==8.14.0",
    "msgpack==1.0.6",
    "networkx==3.1",
    "nodeenv==1.8.0",
    "packaging==23.1",
    "pandas>=2.3.3",
    "pandas-gbq>=0.30.0",
    "parsedatetime==2.6",
    "pathspec==0.11.2",
    "platformdirs==3.11.0",
    "pluggy==1.3.0",
    "pre-commit==3.4.0",
    "proto-plus==1.22.3",
    "protobuf==4.24.3",
    "pyasn1==0.5.0",
    "pyasn1-modules==0.3.0",
    "pycparser==2.21",
    "pydantic==1.10.13",
    "pygments==2.16.1",
    "pyrsistent==0.19.3",
    "pytest==7.4.2",
    "python-dateutil==2.8.2",
    "python-slugify==8.0.1",
    "pytimeparse==1.1.8",
    "pytz==2023.3.post1",
    "pyyaml==6.0.1",
    "regex==2023.8.8",
    "requests==2.32.2",
    "rich==13.6.0",
    "rsa==4.9",
    "ruamel-yaml==0.17.33",
    "ruamel-yaml-clib==0.2.7",
    "six==1.16.0",
    "sqlfluff==2.3.2",
    "sqlfluff-templater-dbt==2.3.2",
    "sqlparse==0.5.0",
    "tblib==2.0.0",
    "text-unidecode==1.3",
    "tqdm==4.66.3",
    "typing-extensions==4.8.0",
    "urllib3==1.26.19",
    "virtualenv==20.24.5",
    "zipp==3.19.1",
]
